remove use of media9 package
https://www.overleaf.com/learn/latex/Questions/How_can_I_embed_a_video_in_my_PDF_using_LaTeX%3F

It used to be possible via the media9 package but due to Adobe ceasing support for Flash at the end of 2020, versions of the media9 package which rely on Flash should no longer be used to embed video or other multimedia content into LaTeX-generated PDF files, including any projects created on Overleaf. However, it is likely that the media9 package will be updated by its author to use non-Flash-based methods for embedding video into LaTeX-generated PDFs.